requirements.txt
pytest
pytest-playwright
psycopg2-binary
